+++ web-frontend/modules/core/plugins/realTimeHandler.js
@@ -230,6 +230,13 @@ export class RealTimeHandler {
         store.dispatch('application/forceDelete', application)
       }
     })
+
+    this.registerEvent('applications_reordered', ({ store }, data) => {
+      const group = store.getters['group/get'](data.group_id)
+      if (group !== undefined) {
+        store.commit('application/ORDER_ITEMS', { group, order: data.order })
+      }
+    })
   }
 }