+++ changelog.md
@@ -7,6 +7,7 @@
 * Made it possible to import a JSON file when creating a table.
 * Made it possible to order the views by drag and drop.
 * Made it possible to order the groups by drag and drop.
+* Made it possible to order the applications by drag and drop.
 * Made it possible to order the tables by drag and drop.
 
 ## Released (2021-05-11)